Swaggified the edit button
Alright so I made the edit button have a black rounded box and when a
user is in edit mode it shows a huge box explaining it's edit mode.

In the future, I'm going to make a save button there which saves
everything. No need to make individual save buttons